+++ CHANGELOG.md
@@ -14,6 +14,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 
+## [1.1.1] - 2026-09-07
+
+### Added
+
+### Changed
+
+### Removed
+
+
 ## [1.1.0] - 2026-09-07
 
 ### Added